LayoutTests/fast/mediastream/captureStream/canvas-rvfc-metadata.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
    <body>
        <canvas id="canvas" width=100 height=100></canvas>
        <video id="video" autoplay width=100 height=100></video>
        <script src="../../../resources/testharness.js"></script>
        <script src="../../../resources/testharnessreport.js"></script>
        <script>

const canvas = document.getElementById("canvas");
const video  = document.getElementById("video");

function drawFrame(color) {
    const ctx = canvas.getContext("2d");
    ctx.fillStyle = color;
    ctx.fillRect(0, 0, canvas.width, canvas.height);
}

function waitForFrame(v) {
    return new Promise((resolve, reject) => {
        const id = v.requestVideoFrameCallback((now, metadata) => resolve({ now, metadata }));
        setTimeout(() => {
            v.cancelVideoFrameCallback(id);
            reject("requestVideoFrameCallback timed out after 5s");
        }, 5000);
    });
}

promise_test(async test => {
    const stream = canvas.captureStream();
    video.srcObject = stream;

    // Let the pipeline initialize before registering the first callback.
    await new Promise(resolve => requestAnimationFrame(() => setTimeout(resolve, 0)));

    drawFrame("green");
    const { metadata: meta1 } = await waitForFrame(video);

    // Dimensions must match the canvas.
    assert_equals(meta1.width,  canvas.width,  "width matches canvas");
    assert_equals(meta1.height, canvas.height, "height matches canvas");

    // presentedFrames must be at least 1.
    assert_greater_than(meta1.presentedFrames, 0, "presentedFrames > 0");

    // presentationTime is a DOMHighResTimeStamp (ms since time origin), must be positive.
    assert_greater_than(meta1.presentationTime, 0, "presentationTime > 0");

    // expectedDisplayTime must be at or after presentationTime.
    assert_greater_than_equal(meta1.expectedDisplayTime, meta1.presentationTime,
        "expectedDisplayTime >= presentationTime");

    // mediaTime is stream-relative (seconds since stream start), must be non-negative
    // and well below the system uptime, confirming it is not the raw clock value.
    assert_greater_than_equal(meta1.mediaTime, 0, "mediaTime >= 0");
    assert_less_than(meta1.mediaTime, 60, "mediaTime is stream-relative, not an absolute clock value");

    // captureTime must be provided for canvas capture frames.
    assert_not_equals(meta1.captureTime, undefined, "captureTime is set");
    assert_greater_than(meta1.captureTime, 0, "captureTime > 0");

    // Draw a second frame and verify all monotonically-increasing fields advance.
    drawFrame("green");
    const { metadata: meta2 } = await waitForFrame(video);

    assert_greater_than(meta2.presentedFrames, meta1.presentedFrames, "presentedFrames increases");
    assert_greater_than(meta2.presentationTime, meta1.presentationTime, "presentationTime increases");
    assert_greater_than(meta2.mediaTime, meta1.mediaTime, "mediaTime increases");

}, "requestVideoFrameCallback metadata for canvas captureStream (2d context)");

        </script>
    </body>
</html>
